+++ include/metashell/pragma_ls.hpp
@@ -1,5 +1,5 @@
-#ifndef METASHELL_INCLUDE_TYPE_HPP
-#define METASHELL_INCLUDE_TYPE_HPP
+#ifndef METASHELL_PRAGMA_LS_HPP
+#define METASHELL_PRAGMA_LS_HPP
 
 // Metashell - Interactive C++ template metaprogramming shell
 // Copyright (C) 2016, Abel Sinkovics (abel@sinkovics.hu)
@@ -17,25 +17,35 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-#include <boost/filesystem/path.hpp>
+#include <metashell/iface/pragma_handler.hpp>
 
-#include <string>
+#include <boost/operators.hpp>
+
+#include <iosfwd>
 
 namespace metashell
 {
-  namespace data
-  {
-    enum class include_type
-    {
-      sys = 0,
-      quote = 1
-    };
+  class shell;
 
-    std::string to_string(include_type type_);
-    std::string include_code(include_type type_,
-                             const boost::filesystem::path& path_);
-    std::string include_dotdotdot(include_type type_);
-  }
+  class pragma_ls : public iface::pragma_handler
+  {
+  public:
+    explicit pragma_ls(shell& shell_);
+
+    virtual iface::pragma_handler* clone() const override;
+
+    virtual std::string arguments() const override;
+    virtual std::string description() const override;
+
+    virtual void run(const data::command::iterator& name_begin_,
+                     const data::command::iterator& name_end_,
+                     const data::command::iterator& args_begin_,
+                     const data::command::iterator& args_end_,
+                     iface::displayer& displayer_) const override;
+
+  private:
+    shell& _shell;
+  };
 }
 
 #endif