+++ test/system/app/pp/test_includes.cpp
@@ -18,6 +18,7 @@
 #include <metashell_system_test/json_generator.hpp>
 #include <metashell_system_test/run_metashell.hpp>
 #include <metashell_system_test/system_test_config.hpp>
+#include <metashell_system_test/util.hpp>
 
 #include <boost/range/adaptors.hpp>
 #include <boost/algorithm/string/predicate.hpp>
@@ -32,19 +33,6 @@ using namespace metashell_system_test;
 
 namespace
 {
-  boost::optional<std::string> remove_prefix(const std::string& prefix_,
-                                             const std::string& s_)
-  {
-    if (boost::algorithm::starts_with(s_, prefix_))
-    {
-      return s_.substr(prefix_.size());
-    }
-    else
-    {
-      return boost::none;
-    }
-  }
-
   bool metashell_standard_header_path(const boost::filesystem::path& path_)
   {
     const auto ends_with = [&path_](std::vector<std::string> suffix_)
@@ -76,7 +64,7 @@ namespace
                find(system_test_config::metashell_args().begin(), end, "--"),
                end, [&path_](const std::string& s_)
                {
-                 const auto path = remove_prefix("-I", s_);
+                 const auto path = try_to_remove_prefix("-I", s_);
                  return path && *path == path_;
                }) != end;
   }